--- src/main/java/io/github/janmalch/kino/entity/Specification.java
@@ -1,8 +1,0 @@
-package io.github.janmalch.kino.entity;
-
-import javax.persistence.Query;
-
-public interface Specification {
-
-  Query toQuery();
-}
