+++ src/main/java/io/github/janmalch/kino/repository/specification/Specification.java
@@ -1,0 +1,8 @@
+package io.github.janmalch.kino.repository.specification;
+
+import javax.persistence.TypedQuery;
+
+public interface Specification<T> {
+
+  TypedQuery<T> toQuery();
+}
